DEPLOY-8-END-PROJECT

+++ src/components/comments/Comments.jsx
@@ -2,6 +2,7 @@ import React, { useEffect } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { useNavigate } from 'react-router-dom';
 import { getSearchResponse, sendRequestComment, sendRequestDeleteComment } from '../../redux/comments/commentAction';
+import swal from 'sweetalert';
 
 const Comments = () => {
 
@@ -21,7 +22,21 @@ const Comments = () => {
 
 
     const handleDelete = (u)=>{
-        dispatch(sendRequestDeleteComment(u.id));
+        swal({
+            title: "حذف کاربر !",
+            text: `آیا از حذف آیتم ${u.id} اطمینان دارید ؟`,
+            icon: "warning",
+            buttons: ['لغو' , 'تایید'],
+            dangerMode: true,
+          })
+          .then((willDelete) => {
+            if (willDelete) {
+                dispatch(sendRequestDeleteComment(u.id));
+            }
+            else {
+              swal("عملیات لغو شد .");
+            }
+        });
     }
